refactorize
