+++ src/cpp_core/CMakeLists.txt
@@ -1,0 +1,18 @@
+find_package(Boost REQUIRED COMPONENTS graph)
+message(STATUS "Boost Libaries: ${Boost_LIBRARIES} ${Boost_INCLUDE_DIRS}")
+
+add_library(gs_core SHARED
+  cc_and_walk.cxx
+)
+set_target_properties(
+  gs_core PROPERTIES
+  LINKER_LANGUAGE CXX
+  CXX_STANDARD 20
+)
+target_include_directories(gs_core PUBLIC include)
+target_link_libraries(gs_core PUBLIC Boost::graph)
+
+install(TARGETS gs_core
+  LIBRARY DESTINATION lib
+  PUBLIC_HEADER DESTINATION include
+)
