+++ src/cpp_pybind11/CMakeLists.txt
@@ -2,5 +2,6 @@ set(PYBIND11_NEWPYTHON ON)
 find_package(pybind11 CONFIG REQUIRED)
 
 pybind11_add_module(_core walk_through_binding.cxx)
+target_link_libraries(_core PRIVATE gs_core)
 
 install(TARGETS _core LIBRARY DESTINATION graph_segment)
