+++ src/cpp_pybind11/walk_through_binding.cxx
@@ -4,19 +4,11 @@
 #include <boost/graph/adjacency_list.hpp>
 #include <vector>
 
-// Define your Graph type in Boost
-using Graph = boost::adjacency_list<boost::vecS, boost::vecS, boost::undirectedS>;
-
-// The original function
-std::vector<std::vector<int>> get_tracks(const Graph &G, double cc_cut, double th_min, double th_add) {
-    // Your implementation or library code here
-    // Example placeholder implementation
-    return {{0, 1, 2}, {3, 4}};
-}
+#include "graph_segment/cc_and_walk.h"
 
 // Helper to convert NetworkX graph to Boost Graph
-Graph convert_nx_to_bgl(pybind11::object nx_graph) {
-    Graph G;
+GraphSegment::Graph convert_nx_to_bgl(pybind11::object nx_graph) {
+    GraphSegment::Graph G;
 
     // Add nodes
     pybind11::list nodes = nx_graph.attr("nodes")();
@@ -40,8 +32,8 @@ PYBIND11_MODULE(_core, m) {
     m.doc() = "C++ module for Connected Components and Walk Through";
 
     m.def("get_tracks", [](pybind11::object nx_graph, double cc_cut, double th_min, double th_add) {
-        Graph G = convert_nx_to_bgl(nx_graph);
-        return get_tracks(G, cc_cut, th_min, th_add);
+        auto G = convert_nx_to_bgl(nx_graph);
+        return GraphSegment::get_tracks(G, cc_cut, th_min, th_add);
     }, "Process tracks in a graph",
        pybind11::arg("nx_graph"), pybind11::arg("cc_cut"), pybind11::arg("th_min"), pybind11::arg("th_add"));
 }
